Rename project to ocean_data_qc
Added package fyskem for all fyskemqc related modules that previously
was in the root level.

Part of #46
