+++ src/ocean_data_qc/fyskem/configs/detection_limit_check.yaml
@@ -1,0 +1,30 @@
+ALKY:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.5
+AMON:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.2
+CPHL:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.2
+DOXY_BTL:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.1
+DOXY_CTD:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.2
+H2S:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 4
+NTOT:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 5
+NTRA:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.1
+NTRI:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.02
+NTRZ:
+  global: !!python/object:ocean_data_qc.fyskem.qc_checks.DetectionLimitCheck
+    limit: 0.1
