+++ tests/fyskem/test_fyskemqc.py
@@ -1,8 +1,8 @@
 import pytest
-from fyskemqc.fyskemqc import FysKemQc
-from fyskemqc.qc_flag import QcFlag
-from fyskemqc.qc_flag_tuple import QcField, QcFlagTuple
-from fyskemqc.qc_flags import QcFlags
+from ocean_data_qc.fyskem.qc_flag import QcFlag
+from ocean_data_qc.fyskem.qc_flag_tuple import QcField, QcFlagTuple
+from ocean_data_qc.fyskem.qc_flags import QcFlags
+from ocean_data_qc.fyskemqc import FysKemQc
 
 from tests.setup_methods import generate_data_frame, generate_data_frame_of_length
