+++ tests/fyskem/test_range_qc.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
-from fyskemqc.parameter import Parameter
-from fyskemqc.qc_flag import QcFlag
-from fyskemqc.qc_flag_tuple import QcField
-from fyskemqc.range_qc import RangeQc
+from ocean_data_qc.fyskem.parameter import Parameter
+from ocean_data_qc.fyskem.qc_flag import QcFlag
+from ocean_data_qc.fyskem.qc_flag_tuple import QcField
+from ocean_data_qc.fyskem.range_qc import RangeQc
 
 from tests.setup_methods import generate_data_frame, generate_range_check_configuration
