+++ tests/metadata/test_visit.py
@@ -1,8 +1,8 @@
 import pytest
-from fyskemqc.errors import VisitError
-from fyskemqc.metadata_qc.metadata_flag import MetadataFlag
-from fyskemqc.metadata_qc.metadata_qc_field import MetadataQcField
-from fyskemqc.visit import Visit
+from ocean_data_qc.errors import VisitError
+from ocean_data_qc.metadata.metadata_flag import MetadataFlag
+from ocean_data_qc.metadata.metadata_qc_field import MetadataQcField
+from ocean_data_qc.metadata.visit import Visit
 from setup_methods import generate_data_frame_from_data_list
