+++ tests/metadata/test_wadep_qc.py
@@ -1,8 +1,8 @@
 import pytest
-from fyskemqc.metadata_qc.metadata_flag import MetadataFlag
-from fyskemqc.metadata_qc.metadata_qc_field import MetadataQcField
-from fyskemqc.metadata_qc.wadep_qc import WadepQc
-from fyskemqc.visit import Visit
+from ocean_data_qc.metadata.metadata_flag import MetadataFlag
+from ocean_data_qc.metadata.metadata_qc_field import MetadataQcField
+from ocean_data_qc.metadata.visit import Visit
+from ocean_data_qc.metadata.wadep_qc import WadepQc
 from setup_methods import generate_data_frame_from_data_list
